Now running tests for hypercore, hyperdb, and hyperdrive
