+++ test/hypercore.js
@@ -1,0 +1,2 @@
+var database = require('hypercore')
+require('./common')()
