+++ test/hyperdb.js
@@ -1,0 +1,2 @@
+var database = require('hyperdb')
+require('./common')()
