+++ test/hyperdrive.js
@@ -1,0 +1,2 @@
+var database = require('hyperdrive')
+require('./common')()
